+++ src/roq/binance_futures/gateway.hpp
@@ -94,7 +94,7 @@ struct Gateway final : public server::Handler,
   Rest rest_;
   absl::flat_hash_map<Account, std::unique_ptr<OrderEntry>> order_entry_;
   absl::flat_hash_map<Account, std::unique_ptr<DropCopy>> drop_copy_;
-  std::vector<std::unique_ptr<MarketData>> market_data_;
+  std::vector<std::unique_ptr<MarketData>> market_data_1_, market_data_2_;
   // cache
   std::vector<MBPUpdate> bids_, asks_;
 };