+++ src/roq/binance_futures/market_data.hpp
@@ -41,7 +41,7 @@ struct MarketData final : public web::socket::Client::Handler, public json::Mark
     virtual void operator()(Trace<StatisticsUpdate> const &, bool is_last) = 0;
   };
 
-  MarketData(Handler &, io::Context &, uint16_t stream_id, Shared &, size_t index);
+  MarketData(Handler &, io::Context &, uint16_t stream_id, Priority, Shared &, size_t index);
 
   MarketData(MarketData &&) = delete;
   MarketData(MarketData const &) = delete;
@@ -90,9 +90,10 @@ struct MarketData final : public web::socket::Client::Handler, public json::Mark
  private:
   Handler &handler_;
   // config
-  const uint16_t stream_id_;
-  const std::string name_;
-  const size_t index_;
+  uint16_t const stream_id_;
+  Priority const priority_;
+  std::string const name_;
+  size_t const index_;
   // web socket
   std::unique_ptr<web::socket::Client> connection_;
   // buffers