+++ src/roq/binance_futures/shared.hpp
@@ -59,7 +59,26 @@ struct Shared final {
  public:
   auto &get_mbp() { return mbp.clear(); }
 
-  absl::flat_hash_map<Symbol, core::mbp::Sequencer> mbp_sequencer;
+  struct Instrument final {
+    int64_t tob_last_update_id = {};
+    int64_t mbp_last_update_id = {};
+    core::mbp::Sequencer sequencer;
+
+    bool tob_update(int64_t update_id) {
+      if (update_id < tob_last_update_id)
+        return false;
+      tob_last_update_id = update_id;
+      return true;
+    }
+
+    bool mbp_update(int64_t update_id) {
+      if (update_id < mbp_last_update_id)
+        return false;
+      mbp_last_update_id = update_id;
+      return true;
+    }
+  };
+  absl::flat_hash_map<Symbol, Instrument> instruments;
 
  private:
   server::Dispatcher &dispatcher_;